airflow_docker/requirements.txt
apache-airflow==2.9.1
pandas
matplotlib
seaborn
notebook
apache-airflow-providers-papermill
papermill
sqlalchemy 
mysql-connector-python